server/requirements.txt
openenv-core[core]>=0.2.2
pandas>=2.0.0
numpy>=1.24.0
